+++ src/app/code/community/Shopgate/Cloudapi/Block/Customer/Account.php
@@ -19,6 +19,7 @@
  * @copyright Shopgate Inc
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License, Version 2.0
  */
+
 class Shopgate_Cloudapi_Block_Customer_Account extends Mage_Core_Block_Template
 {
     /**
@@ -45,9 +46,9 @@ class Shopgate_Cloudapi_Block_Customer_Account extends Mage_Core_Block_Template
             return $this->getRequestHelper()->cookieGetValue(
                 Shopgate_Cloudapi_Helper_Request::KEY_SGCLOUD_CALLBACK_DATA
             );
-        } else {
-            return '';
         }
+
+        return '';
     }
 
     /**