+++ src/app/code/community/Shopgate/Cloudapi/Model/Frontend/Observer/CheckoutMessages.php
@@ -20,6 +20,20 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License, Version 2.0
  */
 
-class Shopgate_Cloudapi_Model_Api2_Auth_Rest extends Shopgate_Cloudapi_Model_Api2_Resource
+class Shopgate_Cloudapi_Model_Frontend_Observer_CheckoutMessages
 {
+    /**
+     * Adds the cart messages from our custom session
+     */
+    public function execute()
+    {
+        if (!Mage::helper('shopgate_cloudapi/request')->isShopgateCheckout()) {
+            return;
+        }
+
+        $messages = Mage::getSingleton('shopgate_cloudapi/storage_session')->getMessages(true)->getItems();
+        if (!empty($messages)) {
+            Mage::getSingleton('core/session')->addMessages($messages);
+        }
+    }
 }