+++ CHANGELOG.md
@@ -14,8 +14,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Shell script to update our REST attributes / roles
 - Support of app-only coupons (except when EE Customer Segments is installed)
 - Endpoint for wishlist creation
+- Endpoint for wishlist removal
 - Endpoint for wishlist retrieval
 - Endpoint for wishlist product adding
+- Endpoint for wishlist product removal
 - Endpoint for wishlist product retrieval
 ### Changed
 - Initial installed admin role, it no longer uses all resources